feat(core): change from custom install to stow
Now the installation proccess is much easier AND modular with stow

You cah choose to install only one or more of the "packages" in this
dotfiles: `base`, `gui` and `extras`.

The `README.md` also was modified to reflect this new organization of
files

Also added a new snipped plugin to vim, a new alias from `vi` to `nvim`
and fix the highlight line.

This update might have broken some things (especially the scripts), but
I'll look into it later (all in time).