feat(script): add dynamic dotfile path support
Add now support for the filepath of the dotfiles to be dynamic with the
addition of the `~/.dotfiles.path` file that keeps record of the last
place the `install` script was executed.

Also change some (two) scripts to fix some issues I was having:

The errno was using grep, but now I use ripgrep and remap it to grep. I
know that most people won't do that for the sake of ONE SINGLE MISERABLE
dotfiles, so I just escaped the command so that it ignores my alias and
use grep instead.

The fsize (on the other hand) was having the same error as `opac`, but
as I don't use it as often I just missed it. Now I realized and fixed
the issue.

Nice