Add support and keymaps for screenkey
Now there is a script to toggle screenkey and a keymap on i3's config
files for that. Update `README.md` to reflect this new program.

Also fix (or remove) some other configs from `.bash_alias` for bloated
programs.

Add utf-8 option to nvim config as it seems to be struggling to save
files the **right way**.