[+] Add clangd to README
Clangd (which is used as my lsp for nvim. God knows how much I tried to
make ccls work) was added to the list of favorite programs (because it
is, indeed, a favorite of mine).

Also moved ScreenKey from favorite to bloated favorite because it is not
that necessary (let's face it, I forget to use it constantly) even
though it is still handy and very cool to have around and at my
disposal.