Fix some configs and add new options to bash
Now bash has a few new features including the extended globbing (fun
stuff, trust me)

Nvim spelllangs now working, Ctrl now is both on the ctrl (which it
wasn't before) and the caps-lock (which it already was).

The rest is fonts and some code style choices which were not aligned
with my current one.